# build windows ci target in ReleaseSafe for faster zip extraction