Compiling zigup for windows in Debug mode:
```
extracted archive in 12.00 s
extracted archive in 18.02 s
extracted archive in 14.56 s
```

Compiling zigup for windows in ReleaseSafe mode:
```
extracted archive in 3.28 s
extracted archive in 5.35 s
extracted archive in 4.36 s
```